+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project are documented in this file.
 
+## 1.5.1 - UI + ModBox Links Fixes
+
+- About page now only displays one close button.   
+- Suppressed non-critical errors so they don't appear in chrome. 
+- Moved ModBox link generator to the about page. 
+- Fixed z-index issue with ModBox link generator page. 
+- Updated ModBox link generator to inlcude subreddit, ban duration, and, note type. 
+
 ## 1.5.0 - Update Checker
 
 - Automatically checks for new updates every 6 hours.